requirements.txt
streamlit
langchain
langchain-google-genai
pypdf
faiss-cpu
tiktoken